PYIC-3574 Rename package name for VerifiableCredentialJwtValidator.
